:art: Style: Move mdElementType to types
